requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
pydantic==2.6.1
langchain==0.1.16
langchain-core==0.1.45
langchain-community==0.0.34
langchain-google-genai==1.0.3
pypdf==4.2.0
faiss-cpu==1.8.0
python-dotenv==1.0.1
gunicorn==21.2.0
